refactor: fileStyle 분리
